+++ tests/dummy/app/locales/ru/translations.js
@@ -175,6 +175,10 @@ Ember.$.extend(true, translations, {
               'caption': 'Пример работы с настройками',
               'title': ''
             },
+            'toolbar-custom-buttons-example': {
+              'caption': 'Пользовательские кнопки',
+              'title': ''
+            },
             'on-edit-form': {
               'caption': 'Размещение на форме редактирования',
               'title': ''
@@ -385,6 +389,11 @@ Ember.$.extend(true, translations, {
         },
         'settings-example': {
           'caption': 'Components-examples/flexberry-objectlistview/settings-example'
+        },
+        'toolbar-custom-buttons-example': {
+          'caption': 'Components-examples/flexberry-objectlistview/toolbar-custom-buttons-example',
+          'custom-message': 'Привет!',
+          'custom-button-name': 'Передать привет'
         },
         'on-edit-form': {
           'caption': 'FlexberryObjectlistview пример произвольных данных'